+++ src/disjoint_matrix.c
@@ -12,18 +12,18 @@
 
 #include <stdint.h>
 
-uint32_t get_dm_n_lines(const dataset_t* dataset)
+uint64_t get_dm_n_lines(const dataset_t* dataset)
 {
-	uint32_t n = 0;
+	uint64_t n = 0;
 
-	uint32_t n_classes	  = dataset->n_classes;
-	uint32_t* n_class_obs = dataset->n_observations_per_class;
+	uint64_t n_classes	  = dataset->n_classes;
+	uint64_t* n_class_obs = dataset->n_observations_per_class;
 
-	for (uint32_t i = 0; i < n_classes - 1; i++)
+	for (uint64_t i = 0; i < n_classes - 1; i++)
 	{
-		for (uint32_t j = i + 1; j < n_classes; j++)
+		for (uint64_t j = i + 1; j < n_classes; j++)
 		{
-			n += n_class_obs[i] * n_class_obs[j];
+			n += (uint64_t) n_class_obs[i] * (uint64_t) n_class_obs[j];
 		}
 	}
